requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
Cython==0.29.21
decorator==4.4.2
discord==1.0.1
discord.py==1.6.0
distlib==0.3.1
Django==3.1.7
et-xmlfile==1.0.1
filelock==3.0.12
idna==3.1
importlib-metadata==2.0.0
joblib==1.0.1
multidict==5.1.0
numpy==1.19.3
openpyxl==3.0.7
pandas==1.2.0
Pillow==8.1.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pytesseract==0.3.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.5
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
Telethon==1.19.0
threadpoolctl==2.1.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
virtualenv==20.1.0
yarl==1.6.3
zipp==3.4.0
